Reraise Error in insertRule wrapper to fix styled-components (#62)
* Reraise Error in insertRule wrapper

We are running into an issue with styled-components. The current wrapper of insertRule breaks styled components when there is invalid css. Note that invalid css also includes css that is valid in firefox but not chrome.

In styled-components it is expected that insertRule raises an exception when there is an error and does not increment index in that case.  See the following code:
https://github.com/styled-components/styled-components/blob/147b0e9a1f10786551b13fd27452fcd5c678d5e0/packages/styled-components/src/sheet/Tag.js#L35-L43

The solution here is to raise the original error back that the caller can catch on insertRule's error.

* Update mutation.ts

* Add comment